# api/swagger.yaml
openapi: 3.0.3
info:
  title: Athlete Hub API
  description: |
    REST API for the Athlete Hub sports events platform.

    ## Authentication
    - **Protected routes** require `Authorization: Bearer <session_token>`
    - **Public routes** require no authentication

    ## Base URL
    All endpoints are prefixed with `/api`.
  version: "1.0.0"
  contact:
    name: Athlete Hub Dev
servers:
  - url: /api
    description: Same-origin (development & production)
...
paths:
...
  /admin/spei-settlements:
    get:
      tags: [Admin]
      summary: List SPEI settlement queue (platform-mode Stripe charges)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, sent, cancelled, all]
            default: pending
      responses:
        "200":
          description: Settlements with pending/sent totals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeiSettlementListResponse"
  /admin/spei-settlements/{id}/mark-sent:
    post:
      tags: [Admin]
      summary: Mark a pending SPEI settlement as sent (ops confirmed bank transfer)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                notes: { type: string, maxLength: 500 }
      responses:
        "200":
          description: Settlement marked sent
          content:
            application/json:
              schema:
                type: object
                required: [settlement]
                properties:
                  settlement:
                    $ref: "#/components/schemas/SpeiSettlementRow"
        "400":
          description: Not pending, or CLABE reveal not possible
        "404":
          description: Settlement not found
  /organizer/spei-settlements:
    get:
      tags: [Organizer]
      summary: List own SPEI settlements (pending/sent totals; CLABE last4 only)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, sent, cancelled, all]
            default: pending
      responses:
        "200":
          description: Organizer settlements
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeiSettlementListResponse"
        "403":
          description: Insufficient permissions
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SpeiSettlementRow:
      type: object
      required:
        - id
        - payment_id
        - organizer_id
        - amount_cents
        - currency
        - status
        - created_at
        - updated_at
      properties:
        id: { type: integer }
        payment_id: { type: integer }
        organizer_id: { type: integer }
        organizer_name: { type: string }
        event_id: { type: integer, nullable: true }
        event_title: { type: string, nullable: true }
        amount_cents: { type: integer }
        currency: { type: string }
        status:
          type: string
          enum: [pending, sent, cancelled]
        payout_account_id: { type: integer, nullable: true }
        clabe_last4:
          type: string
          nullable: true
          description: Last 4 of CLABE only — never full CLABE
        bank_name: { type: string, nullable: true }
        marked_sent_at: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        created_at: { type: string }
        updated_at: { type: string }
    SpeiSettlementListResponse:
      type: object
      required: [settlements, pending_total_cents, sent_total_cents]
      properties:
        settlements:
          type: array
          items:
            $ref: "#/components/schemas/SpeiSettlementRow"
        pending_total_cents: { type: integer }
        sent_total_cents: { type: integer }